change names of some metadata fields
