+++ metadata/pkg/query-manager/translator.go
@@ -108,12 +108,12 @@ func getBucketsHavingFilteredObjectsQuery(filterConditions bson.A) bson.E {
 				{constants.NAME, getQualifiedNameForMapBucket(constants.NAME)},
 				{constants.REGION, getQualifiedNameForMapBucket(constants.REGION)},
 				{constants.TYPE, getQualifiedNameForMapBucket(constants.TYPE)},
-				{constants.ACCESS, getQualifiedNameForMapBucket(constants.ACCESS)},
+				{constants.BUCKET_ACL, getQualifiedNameForMapBucket(constants.BUCKET_ACL)},
 				{constants.NUMBER_OF_OBJECTS, getQualifiedNameForMapBucket(constants.NUMBER_OF_OBJECTS)},
 				{constants.NUMBER_OF_FILTERED_OBJECTS, getTotalNumberOfObjects(filterConditions)},
 				{constants.TOTAL_SIZE, getQualifiedNameForMapBucket(constants.TOTAL_SIZE)},
 				{constants.FILTERED_BUCKET_SIZE, getTotalSizeForObjects(filterConditions)},
-				{constants.TAGS, getQualifiedNameForMapBucket(constants.TAGS)},
+				{constants.BUCKET_TAGS, getQualifiedNameForMapBucket(constants.BUCKET_TAGS)},
 				//* asking to filter the objects array based on the object level queries given by user
 				{constants.OBJECTS, getFilteredObjectsQuery(filterConditions)},
 			},
